+++ NEWS.md
@@ -1,5 +1,7 @@
 # modeldata (development version)
 
+ * A new function `two_class_sim()` was added to simulate classification data. Other helper functions were also added. 
+
 # modeldata 0.0.2
 
 * The bivariate dataset was missing, this has been corrected (@mdogucu, #5).